+++ cmd/commands/root.go
@@ -5,17 +5,15 @@ import (
 
 	"github.com/Drafteame/draft/internal/data"
 	"github.com/Drafteame/draft/internal/pkg/log"
-	nixversion "github.com/Drafteame/draft/internal/pkg/version/nix"
 )
 
 var Version = "development"
 
 var rootCmd = &cobra.Command{
-	Use:              "draft <command>",
-	Example:          "draft new:service",
-	Version:          Version,
-	Run:              run,
-	PersistentPreRun: persistentPreRun,
+	Use:     "draft <command>",
+	Example: "draft new:service",
+	Version: Version,
+	Run:     run,
 }
 
 func init() {
@@ -28,10 +26,6 @@ func run(cmd *cobra.Command, _ []string) {
 	if err := cmd.Help(); err != nil {
 		log.Exitf(1, "failed to print help: %s", err.Error())
 	}
-}
-
-func persistentPreRun(_ *cobra.Command, _ []string) {
-	nixversion.CheckNixModulesVersion()
 }
 
 func GetCmd() *cobra.Command {