+++ internal/actions/newdomain/exec.go
@@ -17,7 +17,7 @@ func (nd *NewDomain) exec() error {
 
 	// Only create providers and domain for postgres
 	if nd.input.DBType == "postgres" {
-		creators = append(creators, nd.createProviders, nd.createDomain)
+		creators = append(creators, nd.createDomain, nd.createDomainProviders, nd.createGlobalProviders)
 	}
 
 	for _, creator := range creators {
@@ -35,12 +35,14 @@ func (nd *NewDomain) createAllDirs() error {
 		nd.input.DomainPath + "/repository",
 	}
 
+	// Only create postgres-specific directories
 	if nd.input.DBType == "postgres" {
 		dirList = append(dirList,
 			nd.input.DomainPath+"/providers",
 			nd.input.DomainPath+"/repository/builders",
 			nd.input.DomainPath+"/repository/daos",
 			nd.input.DomainPath+"/domain/options",
+			"pkg/providers/generators/nanoid/tableid",
 		)
 	}
 
@@ -81,6 +83,7 @@ func (nd *NewDomain) createPostgresService() error {
 		{Path: nd.input.DomainPath + "/service/service_test.go", Data: nd.tmpl.Service.ServiceTestGo},
 		{Path: nd.input.DomainPath + "/service/search_one.go", Data: nd.tmpl.Service.SearchOneGo},
 		{Path: nd.input.DomainPath + "/service/search_one_test.go", Data: nd.tmpl.Service.SearchOneTestGo},
+		{Path: nd.input.DomainPath + "/service/provide.go", Data: nd.tmpl.Service.ProvideGo},
 	}
 
 	for _, file := range fileList {
@@ -143,6 +146,7 @@ func (nd *NewDomain) createPostgresRepository() error {
 		{Path: nd.input.DomainPath + "/repository/daos/daos.go", Data: nd.tmpl.Repository.Postgres.Daos.DaosGo},
 		{Path: nd.input.DomainPath + "/repository/daos/delete.go", Data: nd.tmpl.Repository.Postgres.Daos.DeleteGo},
 		{Path: nd.input.DomainPath + "/repository/daos/update.go", Data: nd.tmpl.Repository.Postgres.Daos.UpdateGo},
+		{Path: nd.input.DomainPath + "/repository/provide.go", Data: nd.tmpl.Repository.Postgres.ProvideGo},
 	}
 
 	for _, file := range fileList {
@@ -170,7 +174,7 @@ func (nd *NewDomain) createDynamoRepository() error {
 	return nil
 }
 
-func (nd *NewDomain) createProviders() error {
+func (nd *NewDomain) createDomainProviders() error {
 	fileList := []dtos.FileEntry{
 		{Path: nd.input.DomainPath + "/providers/generator.go", Data: nd.tmpl.Providers.GeneratorGo},
 		{Path: nd.input.DomainPath + "/providers/service.go", Data: nd.tmpl.Providers.ServiceGo},
@@ -182,12 +186,7 @@ func (nd *NewDomain) createProviders() error {
 		}
 	}
 
-	switch nd.input.DBType {
-	case "postgres":
-		return nd.createPostgresProviders()
-	default:
-		return fmt.Errorf("unsupported providers db type: %s", nd.input.DBType)
-	}
+	return nd.createPostgresProviders()
 }
 
 func (nd *NewDomain) createPostgresProviders() error {
@@ -213,6 +212,20 @@ func (nd *NewDomain) createDomain() error {
 		{Path: nd.input.DomainPath + "/domain/options/update_fields.go", Data: nd.tmpl.Domain.Options.UpdateFieldsGo},
 		{Path: nd.input.DomainPath + "/domain/domain.go", Data: nd.tmpl.Domain.DomainGo},
 		{Path: nd.input.DomainPath + "/domain/errors.go", Data: nd.tmpl.Domain.ErrorsGo},
+	}
+
+	for _, file := range fileList {
+		if err := files.Create(file.Path, file.Data); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (nd *NewDomain) createGlobalProviders() error {
+	fileList := []dtos.FileEntry{
+		{Path: "pkg/providers/generators/nanoid/tableid/" + nd.input.DomainNameLower + ".go", Data: nd.tmpl.Providers.GeneratorsNanoidTableid.ProvideGo},
 	}
 
 	for _, file := range fileList {