+++ internal/dtos/update_nix_modules.go
@@ -1,5 +1,0 @@
-package dtos
-
-type UpdateNixModules struct {
-	ShouldUpdateNixModules bool
-}