+++ internal/templates/domains_service.go
@@ -20,6 +20,7 @@ type Service struct {
 	ServiceTestGo   []byte
 	UpdateGo        []byte
 	UpdateTestGo    []byte
+	ProvideGo       []byte
 	Dynamo          ServiceDynamo
 }
 
@@ -47,6 +48,7 @@ func loadDomainsService(v *Service, data any) error {
 				loadServiceServiceTestGo,
 				loadServiceUpdateGo,
 				loadServiceUpdateTestGo,
+				loadServiceProvideGo,
 			)
 		case "dynamo":
 			loaders = append(loaders, loadServiceDynamo)
@@ -272,6 +274,20 @@ func loadServiceUpdateTestGo(v *Service, data any) error {
 	}
 
 	v.UpdateTestGo = content
+
+	return nil
+}
+
+func loadServiceProvideGo(v *Service, data any) error {
+	name := "domains/service/postgres/provide.go"
+	path := "tmpl/domain/service/postgres/provide.go.tmpl"
+
+	content, err := loadTemplate(name, path, data, domain)
+	if err != nil {
+		return err
+	}
+
+	v.ProvideGo = content
 
 	return nil
 }